Re-configure docs to not rely on plugins being bundled. #17
The decision to not bundle plugins with Front Street, in general, out of the box is a bit of a no-brainer. However, one thing to note here with the docs is that I've also decided to include all JavaScript files separately and not bundle it together. If I were making the fastest site possible, I would bundle it all with Webpack, but instead I want it to be clear to non-NPM folks what's included when they view the source of the docs.